Added int cast

+++ mod_mshield.c
@@ -647,7 +647,7 @@ mshield_access_checker(request_rec *r)
     click_json = cJSON_CreateObject();
     cJSON_AddItemToObject(click_json, "uuid", cJSON_CreateString(cr->session->data->uuid));
     cJSON_AddItemToObject(click_json, "url", cJSON_CreateString(cr->r->unparsed_uri));
-    cJSON_AddItemToObject(click_json, "timestamp", cJSON_CreateNumber(cr->r->request_time));
+    cJSON_AddItemToObject(click_json, "timestamp", cJSON_CreateNumber((double)cr->r->request_time));
     cJSON_AddItemToObject(click_json, "uri", cJSON_CreateString(cr->r->uri));
 
     kafka_produce(r->pool, r, &config->kafka, config->kafka.topic_analyse, &config->kafka.rk_topic_analyse, RD_KAFKA_PARTITION_UA, cJSON_Print(click_json));
